updated the software download/compile script to handle the new podaac website

+++ software/update.sh
@@ -1,9 +1,12 @@
 #!/bin/bash -ue
 
+# keeping track of where I am
+DIR_NOW=$(cd $(dirname $BASH_SOURCE); pwd)
+
 #update this as needed
 VERSION="2010-03-31"
-REMOTE_HOST="podaac-w10n.jpl.nasa.gov"
-REMOTE_DIR="w10n/allData/grace/sw"
+REMOTE_HOST=https://podaac-tools.jpl.nasa.gov
+REMOTE_DIR="drive/files/allData/grace/sw"
 FILE_ROOT="GraceReadSW_L1_"
 FILE_EXT=".tar.gz"
 
@@ -11,12 +14,26 @@ FILE_EXT=".tar.gz"
 LOCAL_DIR=$(cd $(dirname $BASH_SOURCE); pwd)
 CWD=$PWD
 FILENAME=$FILE_ROOT$VERSION.tar.gz
-REMOTE_FILE=http://$REMOTE_HOST/$REMOTE_DIR/$FILENAME
+REMOTE_FILE=$REMOTE_HOST/$REMOTE_DIR/$FILENAME
 LOCAL_FILE=$LOCAL_DIR/$FILE_ROOT$VERSION$FILE_EXT
 SW_DIR=$LOCAL_DIR/RELEASE_$VERSION
 
+#need credentials
+SECRETFILE=$DIR_NOW/../secret.txt
+if [ ! -e "$SECRETFILE" ]
+then
+  echo "ERROR: file $SECRETFILE missing: create this file with your PO.DAAC username and password, each in one single line."
+  exit 3
+fi
+USERNAME=$(head -n1 $SECRETFILE)
+PASSWORD=$(tail -n1 $SECRETFILE)
+
 #download (if not already)
-[ ! -e $LOCAL_FILE ] && wget --no-directories $REMOTE_FILE
+[ ! -e $LOCAL_FILE ] && wget \
+  --user=$USERNAME \
+  --password=$PASSWORD \
+  --no-directories \
+  $REMOTE_FILE
 
 [[ ! "${@/--clean/}" == "$@" ]] && rm -fvr $SW_DIR
